pyproject.toml
[tool.poetry]
name = "it-helpdesk-system"
version = "0.1.0"
description = ""
authors = ["mbusireddy <m.busireddy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
langchain = "^0.3.25"
langchain-community = "^0.3.24"
langchain-core = "^0.3.61"
langchain-text-splitters = "^0.3.8"
langgraph = "^0.4.7"
langgraph-checkpoint = "^2.0.26"
langgraph-prebuilt = "^0.2.1"
langgraph-sdk = "^0.1.70"
langsmith = "^0.3.42"
streamlit = "^1.45.1"
chromadb = "^0.5.23"
ollama = "^0.4.8"
sqlalchemy = "^2.0.41"
alembic = "^1.16.1"
pydantic = "^2.11.5"
pydantic-settings = "^2.9.1"
python-multipart = "^0.0.20"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
loguru = "^0.7.3"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"
aiohttp = "^3.11.18"
numpy = "^2.2.6"
pandas = "^2.2.3"
python-dotenv = "^1.1.0"
pyyaml = "^6.0.2"
jinja2 = "^3.1.6"
tenacity = "^9.1.2"
typer = "^0.15.4"
rich = "^14.0.0"
click = "^8.1.8"
bcrypt = "^4.3.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
